+++ src/serve/cache/stores/DatabaseStore.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Joe J. Howard
+ * @license   https://github.com/Serve-Framework/Framework/blob/master/LICENSE
+ */
+
+namespace serve\cache\stores;
+
+/**
+ * Cache database storage.
+ *
+ * @author Joe J. Howard
+ */
+class DatabaseStore implements StoreInterface
+{
+
+}