+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+### 8.1.0 <small>(2022-18-09)</small>
+
+Added new `Graphql` service.
+
+#### New
+
+* `Graphql` service.
+
+#### Changes
+
+* None
+
+#### Improvements
+
+* None
+
+--------------------------------------------------------
 ### 8.0.0 <small>(2022-18-09)</small>
 
 The major version bump is due to dropped support for PHP `7.4` and a several breaking changes.